src/nlp/Chatbot/chain/requirements.txt
requests==2.31.0
simsimd==3.7.7
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pymongo==4.7.0
openai==1.55.3
langchain-openai==0.1.6
langchain-core==0.1.52
langchain-mongodb==0.1.3
langchain==0.1.20
streamlit==1.35.0
pandas==2.2.1
pymupdf==1.24.10